+++ src/lib/pages/proposal-details/components/DepositList.tsx
@@ -1,10 +1,10 @@
-import { Flex, Text } from "@chakra-ui/react";
+import { Flex } from "@chakra-ui/react";
 
 import { useMobile } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
-import { TokenImageRender } from "lib/components/token";
 import type { ProposalDeposit } from "lib/types";
-import { formatUTokenWithPrecision, getTokenLabel } from "lib/utils";
+
+import { DepositAmounts } from "./DepositAmounts";
 
 interface DepositListProps {
   proposalDeposits: ProposalDeposit[];
@@ -27,24 +27,7 @@ export const DepositList = ({ proposalDeposits }: DepositListProps) => {
             type="user_address"
             showCopyOnHover={!isMobile}
           />
-          <div>
-            {deposit.amount.map((token) => (
-              <Flex key={token.denom} alignItems="center" gap={2}>
-                <Text variant="body2">
-                  <Text as="span" fontWeight={700} mr={1}>
-                    {formatUTokenWithPrecision(
-                      token.amount,
-                      token.precision ?? 0,
-                      true,
-                      2
-                    )}
-                  </Text>
-                  {getTokenLabel(token.denom, token.symbol)}
-                </Text>
-                <TokenImageRender logo={token.logo} />
-              </Flex>
-            ))}
-          </div>
+          <DepositAmounts deposit={deposit} />
         </Flex>
       ))}
     </div>